StatusUpdate
I am at the point where i have the secret board and i have the board printing, I also have it so each line is broken up by a line of dashes. But i need to make the actual systems work like so it checks if the colors guessed are in right position and color or right color wrong position and actually give a feedback.